wm: dont manage wine steam floating menu windows
Previously wine steam floating menu windows cannot be removed properly,
they would just stay there (the border is still visible, but the content
is empty).

Now in WindowManager::OnMapRequest(), we will check if the window we're
going to map is...
1. a Wine steam window
2. a dialog window

If the above conditions are satisfied, then we will map this window
directly without adding it to the client tree, in order to prevent the
weird bug.

Currently all of these rules regarding Wine steam floating windows are
hardcoded.

Cookie instance method are also modified to be more atomic.