implement snapshot (incomplete) and bug fixes
* Previously, lots of dummy entries are created in Client::mapper_.
That was because I was using operator[] instead of mapper_.find().
This commit fixes this.
* Remove redundant close() on fstream objects since they are RAII
objects.
* Fix a bug in tree serialization API (when there is only a root node
in the tree, this root node is counted as a leaf and will segfault since
it will try to do c->client()->xxx())
* Snapshot feature isn't complete. Currently it can successfully survive
exceptions and reconstruct Client::mapper_ as well as Client trees, but
it seems that the tree deserialization API is buggy, not sure.