try to remove old snapshot if rename fails
[Reason for removing old snapshot]

The WM will decide whether it should perform error recovery based on the
existence of ~/.cache/Wmderland/snapshot

Normally when the snapshot is successfully loaded, it will be renamed to
snapshot.old. I don't choose to directly remove the file since it contains
useful debugging information.

[The result of being unable to rename/remove old snapshot]

If the corrupted snapshot cannot be renamed or removed, then the WM
might not be able to start next time, since the corrupted snapshot is there
and the WM will constantly try to load it.