Revert commit 5a9bb89
The commit 5a9bb89 was intended to automatically raise
the floating `focused_client` for the user,
but apparently this behavior is not useful when the
`focused_client` is larger than other floating clients,
since other smaller floating clients will be obscured
by the larger floating client.

It would be better to let the user manually switch the focus themselves
by Mod + left_button.